Fire inserted against canNamespace.document
Before firing the "inserted" event we first check that an element is in
the document. In server-side rendering the ownerDocument is usually not
the document the element was inserted into.

Because of this, this change makes it so that we check against
`canNamespace.document`. In SSR scenarios we will use a Zone to set this
property to be the document the element is being inserted into.

Closes #17