Makes attr.set CHECKED work with the boolean condition
This makes sure that boolean attributes work, so that we can do stuff
like:

```js
attr.set(input, "checked");
```

Which will set it to true. You can still explicitly set it to false
with:

```js
attr.set(input, "checked", false);
```

So the change is to make anything that's not `false` be truthy.