Making dev.warn and dev.log noops in production
Instead of the entire dev module being removed in production, the module will now export empty functions for dev.warn and dev.log. This way, if we forget to remove a call to one of these functions in production it will not throw an error for 'cannot call warn of undefined'.