Add support for slim loader
In a slim build, the global `System` is not set and `can-util/js/import/`
will just resolve the import with empty module.

This change makes it so can-import checks for `stealRequire` and calls
`steal.import` (which is transpiled to `stealRequire.dynamic` during the slim
build) instead of resolving right away.